Modify call to `collection.find()` to remove empty fields object
According to the docs:
http://mongodb.github.io/node-mongodb-native/api-generated/collection.html#find

The only time when you would call it with 3 arguments (ignoring the callback) is
`find(selector, fields, options)`, this is not what we are doing so it should
be removed. We only ever specify fields in `options`.

Also the only place where the projection query can be is in the 2nd argument.
http://docs.mongodb.org/manual/reference/operator/projection/meta/#projection
This is necessary to return the `textScore` $meta associated with a full text
search result. For some reason it does not work as an option in the 3rd argument.