Use through2 for Transform stream
Using through meant that if the data was read from the find()
readStream asynchronously, any data events prior to the read()
or pipe() would be lost. This is due to the incompatiblity of
stream backpressure.

Using through2 the data is correctly buffered, so that the data
is ready to read whenever the consumer is ready.